separated set storage
